crew-ai/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "A multi-agent AI system for stock market analysis and investment reporting."
authors = ["Sridhar"]
license = "MIT"

# Define the package location
packages = [{ include = "src" }]
# Define the source for the package
[[tool.poetry.source]]
name = "crewai"
url = "https://pypi.org/simple"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
crewai = ">=0.108.0"
yfinance = ">=0.2.55"
python-dotenv = ">=1.1.0"

# Optional dependencies for development and testing
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
